+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 Changes to Calva.
 
 ## [Unreleased]
+- [Add Babashka Jack-in option](https://github.com/BetterThanTomorrow/calva/issues/1001)
+- [Add nbb Jack-in option](https://github.com/BetterThanTomorrow/calva/issues/1311)
 - Fix: [Not currently possible to Jack-in to a ClojureScript nREPL Server](https://github.com/BetterThanTomorrow/calva/issues/1310)
 - [Update deps.clj version to 0.0.19](https://github.com/BetterThanTomorrow/calva/issues/1319)